+++ src/VaultSharp/V1/SecretsEngines/Database/IDatabaseSecretsEngine.cs
@@ -23,6 +23,6 @@ namespace VaultSharp.V1.SecretsEngines.Database
         /// <returns>
         /// The secret with the <see cref="UsernamePasswordCredentials" /> as the data.
         /// </returns>
-        Task<Secret<UsernamePasswordCredentials>> GenerateCredentialsAsync(string roleName, string mountPoint = SecretsEngineDefaultPaths.Database, string wrapTimeToLive = null);
+        Task<Secret<UsernamePasswordCredentials>> GetCredentialsAsync(string roleName, string mountPoint = SecretsEngineDefaultPaths.Database, string wrapTimeToLive = null);
     }
 }