+++ src/VaultSharp/V1/SecretsEngines/GoogleCloud/IGoogleCloudSecretsEngine.cs
@@ -26,7 +26,7 @@ namespace VaultSharp.V1.SecretsEngines.GoogleCloud
         /// <returns>
         /// The secret with the <see cref="GoogleCloudOAuth2Token" /> as the data.
         /// </returns>
-        Task<Secret<GoogleCloudOAuth2Token>> GenerateOAuth2TokenAsync(string roleset, string mountPoint = SecretsEngineDefaultPaths.GoogleCloud, string wrapTimeToLive = null);
+        Task<Secret<GoogleCloudOAuth2Token>> GetOAuth2TokenAsync(string roleset, string mountPoint = SecretsEngineDefaultPaths.GoogleCloud, string wrapTimeToLive = null);
 
         /// <summary>
         /// Generates a service account key.
@@ -53,6 +53,6 @@ namespace VaultSharp.V1.SecretsEngines.GoogleCloud
         /// <returns>
         /// The secret with the <see cref="GoogleCloudOAuth2Token" /> as the data.
         /// </returns>
-        Task<Secret<GoogleCloudServiceAccountKey>> GenerateServiceAccountKeyAsync(string roleset, ServiceAccountKeyAlgorithm keyAlgorithm = ServiceAccountKeyAlgorithm.KEY_ALG_RSA_2048, ServiceAccountPrivateKeyType privateKeyType = ServiceAccountPrivateKeyType.TYPE_GOOGLE_CREDENTIALS_FILE,  string mountPoint = SecretsEngineDefaultPaths.GoogleCloud, string wrapTimeToLive = null);
+        Task<Secret<GoogleCloudServiceAccountKey>> GetServiceAccountKeyAsync(string roleset, ServiceAccountKeyAlgorithm keyAlgorithm = ServiceAccountKeyAlgorithm.KEY_ALG_RSA_2048, ServiceAccountPrivateKeyType privateKeyType = ServiceAccountPrivateKeyType.TYPE_GOOGLE_CREDENTIALS_FILE,  string mountPoint = SecretsEngineDefaultPaths.GoogleCloud, string wrapTimeToLive = null);
     }
 }