+++ src/VaultSharp/V1/SecretsEngines/SSH/ISSHSecretsEngine.cs
@@ -1,6 +1,32 @@
-﻿namespace VaultSharp.V1.SecretsEngines.SSH
+﻿using System.Threading.Tasks;
+using VaultSharp.V1.Commons;
+
+namespace VaultSharp.V1.SecretsEngines.SSH
 {
-    public class ISSHSecretsEngine
+    /// <summary>
+    /// The SSH Secrets Engine.
+    /// </summary>
+    public interface ISSHSecretsEngine
     {
+        /// <summary>
+        /// Generates a dynamic SSH credentials for a specific username and IP Address based on the named role.
+        /// </summary>
+        /// <param name="roleName"><para>[required]</para>
+        /// Name of the SSH Role.</param>
+        /// <param name="ipAddress"><para>[required]</para>
+        /// The ip address of the remote host.</param>
+        /// <param name="username"><para>[optional]</para>
+        /// The username on the remote host.</param>
+        /// <param name="mountPoint"><para>[optional]</para>
+        /// The mount point for the SSH backend. Defaults to <see cref="SecretsEngineDefaultPaths.SSH" />
+        /// Provide a value only if you have customized the SSH mount point.</param>
+        /// <param name="wrapTimeToLive">
+        /// <para>[required]</para>
+        /// The TTL for the token and can be either an integer number of seconds or a string duration of seconds.
+        /// </param>
+        /// <returns>
+        /// The secret with the SSH credentials.
+        /// </returns>
+        Task<Secret<SSHCredentials>> GetCredentialsAsync(string roleName, string ipAddress, string username = null, string mountPoint = SecretsEngineDefaultPaths.SSH, string wrapTimeToLive = null);
     }
 }