+++ src/VaultSharp/V1/SecretsEngines/SSH/SSHCredentials.cs
@@ -1,0 +1,55 @@
+using Newtonsoft.Json;
+
+namespace VaultSharp.V1.SecretsEngines.SSH
+{
+    /// <summary>
+    /// SSH Credentials.
+    /// </summary>
+    public class SSHCredentials
+    {
+        /// <summary>
+        /// Gets or sets the ip address.
+        /// </summary>
+        /// <value>
+        /// The ip address.
+        /// </value>
+        [JsonProperty("ip")]
+        public string IpAddress { get; set; }
+
+        /// <summary>
+        /// Gets or sets the key.
+        /// </summary>
+        /// <value>
+        /// The key.
+        /// </value>
+        [JsonProperty("key")]
+        public string Key { get; set; }
+
+        /// <summary>
+        /// Type of credentials
+        /// </summary>
+        /// <value>
+        /// The type of the key.
+        /// </value>
+        [JsonProperty("key_type")]
+        public SSHKeyType KeyType { get; set; }
+
+        /// <summary>
+        /// Gets or sets the port.
+        /// </summary>
+        /// <value>
+        /// The port.
+        /// </value>
+        [JsonProperty("port")]
+        public int Port { get; set; }
+
+        /// <summary>
+        /// Gets or sets the username.
+        /// </summary>
+        /// <value>
+        /// The key.
+        /// </value>
+        [JsonProperty("username")]
+        public string Username { get; set; }
+    }
+}