+++ src/VaultSharp/V1/SystemBackend/MFA/Duo/DuoProvider.cs
@@ -3,7 +3,7 @@ using VaultSharp.V1.Commons;
 
 namespace VaultSharp.V1.SystemBackend.MFA.Duo
 {
-    internal class DuoProvider : MFAProviderBase<DuoConfig>, IDuo
+    internal class DuoProvider : AbstractMFAProviderBase<DuoConfig>, IDuo
     {
         public DuoProvider(Polymath polymath) : base(polymath)
         {