+++ src/VaultSharp/V1/SystemBackend/MFA/PingID/PingIDProvider.cs
@@ -3,7 +3,7 @@ using VaultSharp.V1.Commons;
 
 namespace VaultSharp.V1.SystemBackend.MFA.PingID
 {
-    internal class PingIDProvider : MFAProviderBase<PingIDConfig>, IPingID
+    internal class PingIDProvider : AbstractMFAProviderBase<PingIDConfig>, IPingID
     {
         public PingIDProvider(Polymath polymath) : base(polymath)
         {