+++ CHANGELOG.md
@@ -17,6 +17,7 @@
   * GoogleCloud Secrets Engine: Add support for generating OAuth2 Token & Service Account Key.
   * Nomad Secrets Engine: Add support for generating dynamic credentials.
   * RabbitMQ Secrets Engine: Add support for generating dynamic credentials.
+  * SSH Secrets Engine: Add support for generating dynamic credentials.
 
 ## 0.10.4001 (August 10, 2018)